having issues refactoring
