+++ lib/night_writer.rb
@@ -1,9 +1,62 @@
-require_relative 'convert_from_braille'
-require_relative 'convert_to_braille'
+require_relative 'alphabet'
 
 class NightWriter
 
-  ConvertToBraille.encode(plain)
-  ConvertFromBraille.encode(braille)
+  def initialize
+    @alphabet = Alphabet.new
+  end
+
+  def lookup(character, position)
+    @alphabet.braille_letter_hash[character].chars[position]
+  end
+
+  def encode_to_braille(plain)
+    output = []
+    [0,2,4].each do |offset|
+      plain.chars.each do |letter|
+        if letter == letter.upcase
+          output << lookup(:capitalize, offset) << lookup(:capitalize, offset + 1)
+          letter = letter.downcase
+        end
+        output << lookup(letter, offset) << lookup(letter, offset + 1)
+      end
+      output << "\n"
+    end
+    encoded = output.join
+  end
+
+  def encode_from_braille(braille)
+    lines = braille.split("\n")
+    n = lines[0].length
+    m = 3
+    as_one_line = lines.join
+    output = []
+
+    (0..(n-1)).each_slice(2) do |column_offset|
+
+      braille_character = []
+      (0..(m-1)).each do |row_offset|
+        braille_character << as_one_line[(row_offset * n) + column_offset[0]]
+        braille_character << as_one_line[(row_offset * n) + column_offset[1]]
+      end
+
+      decoded_braille = @alphabet.braille_letter_hash.key(braille_character.join)
+
+      def should_capitalize_next(decoded_braille)
+
+      if decoded_braille == :capitalize
+        return true
+      elsif should_capitalize_next
+        output << decoded_braille.upcase
+        return false
+      else
+        output << decoded_braille
+        return false
+      end
+      end
+    end
+    output.join
+  end
 end
+
